fix(cli): migrate pending-only interrupted first pairings; split test pack
Codex P2 x2 (round 5): (1) `setup --service` after an interrupted FIRST
pairing found no current credential and no-opped — interactive resume then
promoted the pending credential back into the desktop keyring before the
service force applied. The migration now moves a pending-only state to the
file backend and commits the marker (the same durable state an explicit
file-mode pairing leaves), so resume finishes file-backed. (2) split the
490-line regression file into keyring_locked_ux_test.go (locked-UX surface)
and device_credential_migration_test.go (backend-flip/migration) per the
<~400 LOC guideline.
